Reorganised structure
Now have two blueprints, need to do main
