+++ app/__init__.py
@@ -14,10 +14,16 @@ app.config.from_object(Config)
 db = SQLAlchemy(app)
 migrate = Migrate(app, db)
 login = LoginManager(app)
-login.login_view = 'login'
+login.login_view = 'auth.login'
+login.login_message = 'Please log in to access this page'
 bootstrap = Bootstrap(app)
 mail = SendGrid(app)
 
+from app.errors import bp as errors_bp
+app.register_blueprint(errors_bp)
+
+from app.auth import bp as auth_bp
+app.register_blueprint(auth_bp, url_prefix='/auth')
 
 from app import routes, models, errors
