+++ app/auth/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+
+bp = Blueprint('auth', __name__)
+
+from app.auth import routes
