+++ app/errors/handlers.py
@@ -1,13 +1,14 @@
 from flask import render_template
 from app import app, db
+from app.errors import bp
 
 
 @app.errorhandler(404)
 def not_found_error(error):
-    return render_template('404.html'), 404
+    return render_template('errors/404.html'), 404
 
 
 @app.errorhandler(500)
 def internal_error(error):
     db.session.rollback()
-    return render_template('500.html'), 500
+    return render_template('errors/500.html'), 500
