+++ app/routes.py
@@ -1,13 +1,10 @@
-from flask_login import current_user, login_user
-from werkzeug.urls import url_parse
+from flask_login import current_user
 from app import app, db
-from flask import render_template, flash, redirect, url_for, request, abort
-from app.forms import LoginForm, DepartmentalRoleForm, RegistrationForm, RegisterAsForm, PreferencesForm, \
-    ResetPasswordRequestForm
+from flask import render_template, flash, redirect, url_for
+from app.forms import DepartmentalRoleForm, PreferencesForm
 from app.models import User, Candidate, Preferences
-from flask_login import logout_user, login_required
+from flask_login import login_required
 import datetime as dt
-from .email import send_email_confirmation, send_password_reset_email
 
 
 @app.route('/')
@@ -26,24 +23,6 @@ def index():
     return render_template('index.html', title="Home", action=action)
 
 
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.username.data).first()
-        if user is None or not user.check_password(form.password.data):
-            flash("Invalid username or password")
-            return redirect(url_for('login'))
-        login_user(user, remember=form.remember_me.data)
-        next_page = request.args.get('next')
-        if not next_page or url_parse(next_page).netloc != '':
-            next_page = url_for('index')
-        return redirect(next_page)
-    return render_template('login.html', title='Sign in', form=form)
-
-
 @app.route('/submit-role', methods=['GET', 'POST'])
 def submit_role():
     form = DepartmentalRoleForm()
@@ -51,48 +30,6 @@ def submit_role():
         flash('Role {} submitted'.format(form.title))
         return redirect(url_for('index'))
     return render_template('submit-role.html', title="Submit a role", form=form)
-
-
-@app.route('/logout')
-def logout():
-    logout_user()
-    return redirect(url_for('index'))
-
-
-@app.route('/register', methods=['POST', 'GET'])
-def register():
-    """passes user through to correct registration form.
-    """
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = RegisterAsForm()
-    if form.validate_on_submit():
-        if form.type_of_user.data == 'cohort leader':
-            abort(404)
-        elif form.type_of_user.data == 'activity manager':
-            redirect_url = 'register_as_activity_manager'
-        else:
-            redirect_url = 'register_as_candidate'
-        return redirect(url_for(redirect_url))
-
-    return render_template('register.html', title='What kid of user are you?', form=form)
-
-
-@app.route('/register-as-candidate', methods=['GET', 'POST'])
-def register_as_candidate():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        candidate = Candidate()
-        form.populate_obj(candidate)
-        candidate.set_password(form.password.data)
-        db.session.add(candidate)
-        db.session.commit()
-        send_email_confirmation(candidate)
-        flash("Congratulations, you've registered successfully. Now check your email to confirm your account")
-        return redirect(url_for('login'))
-    return render_template('register-as-candidate.html', title='Register as a candidate', form=form)
 
 
 @app.route('/profile')
@@ -150,28 +87,4 @@ def my_cohort():
     return render_template('my-cohort.html', data=cohort)
 
 
-@app.route('/reset-password-request', methods=['POST', 'GET'])
-def reset_password_request():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = ResetPasswordRequestForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user:
-            send_password_reset_email(user)
-        flash('Check your email for the instructions to reset your password')
-        return redirect(url_for('login'))
-    return render_template('reset-password-request.html', title='Reset Password', form=form)
-
-
-@app.route('/confirm-email/<token>/')
-@login_required
-def confirm_email(token):
-    if current_user.confirmed:
-        return redirect(url_for('index'))
-    elif current_user.confirm_user(token):
-        flash("Thank you for confirming your account!")
-    else:
-        flash("The link is invalid or expired")
-    return redirect(url_for('index'))
