+++ app/templates/auth/register-as-candidate.html
