+++ app/templates/errors/404.html
@@ -2,7 +2,7 @@
 {% block app_content %}
     <h1>This is awkward</h1>
     <p>
-        It looks like you've requested a page that no longer existed, never existed in the first place, or exists
+        It looks like you've requested a page that no longer exists, never existed in the first place, or exists
         somewhere else. Take another look at the URL or click below to go back to the previous page
     </p>
     <p><a class="btn btn-primary"  href="{{ url_for('index') }}" role="button">Back</a></p>
